Ignore outputs for earlier commands
when a new command is received.

Earlier, due to the use of promises, results of earlier commands were possible to be relayed to the UI/coordinator, which is not part of the spec. In this version, we ensure that responses to earlier commands are dropped (from stdout, stderr) on receipt of a new valid command.